UI: Refactor auth controller so it does less (#27710)
* move some auth controller logic to route page component

* remove unused vars

* fix action handling so this context is retained

* rename authpage to auth-form-page

* rename auth-route-page to auth-splash-page

* link jira VAULT-28251

* wowww typo

* add padding to mfa form alert message

* update component name in tests

* alphabetize args

* use auth helpers for login method

* remove async, await

* rename components

* update jsdoc

* add comment
