+++ ui/tests/acceptance/access/namespaces/index-test.js
@@ -3,24 +3,23 @@
  * SPDX-License-Identifier: BUSL-1.1
  */
 
-import { currentRouteName } from '@ember/test-helpers';
+import { currentRouteName, visit } from '@ember/test-helpers';
 import { module, test } from 'qunit';
 import { setupApplicationTest } from 'ember-qunit';
 import { setupMirage } from 'ember-cli-mirage/test-support';
-import page from 'vault/tests/pages/access/namespaces/index';
-import authPage from 'vault/tests/pages/auth';
+import { login } from 'vault/tests/helpers/auth/auth-helpers';
 
 module('Acceptance | Enterprise | /access/namespaces', function (hooks) {
   setupApplicationTest(hooks);
   setupMirage(hooks);
 
   hooks.beforeEach(function () {
-    return authPage.login();
+    return login();
   });
 
   test('it navigates to namespaces page', async function (assert) {
     assert.expect(1);
-    await page.visit();
+    await visit('/vault/access/namespaces');
     assert.strictEqual(
       currentRouteName(),
       'vault.cluster.access.namespaces.index',
@@ -30,7 +29,7 @@ module('Acceptance | Enterprise | /access/namespaces', function (hooks) {
 
   test('it should render correct number of namespaces', async function (assert) {
     assert.expect(3);
-    await page.visit();
+    await visit('/vault/access/namespaces');
     const store = this.owner.lookup('service:store');
     // Default page size is 15
     assert.strictEqual(store.peekAll('namespace').length, 15, 'Store has 15 namespaces records');
