--- ui/tests/pages/access/namespaces/index.js
@@ -1,10 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: BUSL-1.1
- */
-
-import { create, visitable } from 'ember-cli-page-object';
-
-export default create({
-  visit: visitable('/vault/access/namespaces'),
-});
